fix(auth): require an explicit consent click; extract consent module
Remove the script that auto-clicked the approve anchor on load. A crafted
/authorize URL with an attacker-controlled loopback redirect would otherwise
deliver the code to a listening local process before the user could perceive
the destination, defeating the spec's redirect-host display requirement. The
page now shows the redirect host, client_id, scope, and loopback warning, and
the user clicks Authorise to proceed (the anchor still carries the full
target so the code and state survive).

Move the page rendering and escaping into a dedicated consent module, set the
content-type via HeaderValue::from_static, and fold the triplicated
LOOPBACK_HOSTS constant into a single oauth::common definition shared by the
redirect matcher, the DCR validator, and the consent warning.
